Added module exporter/common. Moved DelegatingXMLStreamWriter and IndentingXMLStreamWriter from specobject export and aspec reporter.
