+++ exporter/common/src/main/java/org/itsallcode/openfasttrace/exporter/common/IndentingXMLStreamWriter.java
@@ -1,4 +1,4 @@
-package org.itsallcode.openfasttrace.report.aspec;
+package org.itsallcode.openfasttrace.exporter.common;
 
 /*-
  * #%L
@@ -10,12 +10,12 @@ package org.itsallcode.openfasttrace.report.aspec;
  * it under the terms of the GNU General Public License as
  * published by the Free Software Foundation, either version 3 of the
  * License, or (at your option) any later version.
- *
+ * 
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- *
+ * 
  * You should have received a copy of the GNU General Public
  * License along with this program.  If not, see
  * <http://www.gnu.org/licenses/gpl-3.0.html>.
@@ -27,7 +27,7 @@ import javax.xml.stream.XMLStreamWriter;
 import java.util.Deque;
 import java.util.LinkedList;
 
-class IndentingXMLStreamWriter extends DelegatingXMLStreamWriter implements AutoCloseable
+public class IndentingXMLStreamWriter extends DelegatingXMLStreamWriter implements AutoCloseable
 {
     private enum State
     {
