+++ exporter/exporter_common/pom.xml
@@ -2,8 +2,8 @@
     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
     xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <artifactId>openfasttrace-reporter-aspec</artifactId>
-    <name>OpenFastTrace augmented specobject Reporter</name>
+    <artifactId>openfasttrace-exporter-common</artifactId>
+    <name>Common helper classes used by multiple exporters</name>
     <parent>
         <relativePath>../../parent/pom.xml</relativePath>
         <groupId>org.itsallcode.openfasttrace</groupId>
@@ -14,10 +14,6 @@
      <dependency>
             <groupId>org.itsallcode.openfasttrace</groupId>
             <artifactId>openfasttrace-api</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.itsallcode.openfasttrace</groupId>
-            <artifactId>openfasttrace-exporter-common</artifactId>
         </dependency>
         <dependency>
             <groupId>org.itsallcode.openfasttrace</groupId>
